[Refactor] (test) Move the common modules to another single sub-package.
