+++ test/integration_test/_test_utils/_instance_value.py
@@ -3,7 +3,7 @@ from smoothcrawler_cluster.crawler import ZookeeperCrawler
 from typing import List
 import json
 
-from .._values import (
+from test._values import (
     # GroupState
     _Runner_Crawler_Value, _Backup_Crawler_Value,
     # common functions
